Break down Models namespace into Models.Exports and Models.Internal
