+++ SMB3Explorer/Models/Exports/BattingSeasonStatistic.cs
@@ -2,11 +2,12 @@
 using CsvHelper.Configuration.Attributes;
 using SMB3Explorer.Enums;
 using SMB3Explorer.Utils;
+
 // ReSharper disable UnusedAutoPropertyAccessor.Global
 // ReSharper disable MemberCanBePrivate.Global
 // ReSharper disable UnusedMember.Global
 
-namespace SMB3Explorer.Models;
+namespace SMB3Explorer.Models.Exports;
 
 public class BattingSeasonStatistic
 {
