+++ SMB3Explorer/Models/Exports/CareerPitchingStatistic.cs
@@ -1,10 +1,11 @@
 ﻿using CsvHelper.Configuration.Attributes;
 using SMB3Explorer.Enums;
 using SMB3Explorer.Utils;
+
 // ReSharper disable UnusedAutoPropertyAccessor.Global
 // ReSharper disable UnusedMember.Global
 
-namespace SMB3Explorer.Models;
+namespace SMB3Explorer.Models.Exports;
 
 public class CareerPitchingStatistic : CareerStatistic
 {
