+++ SMB3Explorer/Models/Exports/FranchisePlayoffStanding.cs
@@ -1,10 +1,11 @@
 ﻿using CsvHelper.Configuration.Attributes;
+
 // ReSharper disable UnusedAutoPropertyAccessor.Global
 // ReSharper disable UnusedMember.Global
 
-namespace SMB3Explorer.Models;
+namespace SMB3Explorer.Models.Exports;
 
-public class FranchiseSeasonStanding
+public class FranchisePlayoffStanding
 {
     [Name("index"), Index(0)]
     public int Index { get; set; }
@@ -38,10 +39,4 @@ public class FranchiseSeasonStanding
     
     [Name("run_differential"), Index(10)]
     public int RunDifferential { get; set; }
-    
-    [Name("win_percentage"), Index(11)]
-    public double WinPercentage { get; set; }
-    
-    [Name("games_back"), Index(12)]
-    public double GamesBack { get; set; }
 }
