+++ SMB3Explorer/Services/ApplicationContext/ApplicationContext.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.ComponentModel;
 using System.Runtime.CompilerServices;
-using SMB3Explorer.Models;
+using SMB3Explorer.Models.Internal;
 
 namespace SMB3Explorer.Services.ApplicationContext;
