+++ SMB3Explorer/Services/DataService/DataServiceFranchiseSeasons.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using Microsoft.Data.Sqlite;
-using SMB3Explorer.Models;
+using SMB3Explorer.Models.Exports;
 using SMB3Explorer.Utils;
 
 namespace SMB3Explorer.Services.DataService;
